books_site/requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.2
daphne==4.1.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.7
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-mathfilters==1.0.0
django-recaptcha==4.0.0
django-taggit==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
executing==2.0.1
flower==2.0.1
h11==0.14.0
huey==2.5.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.20.0
jedi==0.19.1
kombu==5.3.5
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
msgpack==1.0.7
oauthlib==3.2.2
outcome==1.3.0.post0
parso==0.8.3
Pillow==10.1.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.18.1
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.3
stripe==8.2.0
tornado==6.4
traitlets==5.14.1
trio==0.24.0
trio-websocket==0.11.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
zope.interface==6.2
